Added TxtParser, added loading data file functionality

+++ iadip/iadip/ClusterData.cs
@@ -9,7 +9,7 @@ namespace iadip
     class ClusterData
     {
         public double Cost;
-        public double AreaSize;
+        public int AreaSize;
         public int RoomsCount;
         public int BathroomsCount;
     }
